Match float display precision to the original JS engine
- Float REPR: round to RoundPrecision ( default 15 ) significant digits
  like the JS engine's Round(), then print the shortest round-trip form
  via to_chars ( was: fixed 6 decimals )
- Read: parse numbers with stod, not stof — the float-precision parse
  was masked by the 6-digit display ( 3.14 read as 3.1400001... )
- toJSON: shortest round-trip form, unrounded ( JS JSON.stringify behavior )
- Expose SetRoundPrecision to JS; index.html wires the existing
  Round precision field ( invalid input falls back to 15 )
- MathTest: atan2 expectation was calibrated against the stof parse;
  recalibrate to the correct double value

+++ C++/Read.cpp
@@ -212,11 +212,11 @@ Read( iReader& R, char32_t terminator ) {
 				}
 				$.push_back( R.Read() );
 			}
-			if( dotRead ) return MS<Float >( stof( string_Us( $ ) ) );
+			if( dotRead ) return MS<Float >( stod( string_Us( $ ) ) );
 			try {
 				return MS<Bits	>( stol( string_Us( $ ) ) );
 			} catch( out_of_range const& ) {
-				return MS<Float >( stof( string_Us( $ ) ) );
+				return MS<Float >( stod( string_Us( $ ) ) );
 			}
 		}
 		switch ( _ ) {
